character velocity changes and dash states implemented
Dash state added,
Mobility manager refactored to no longer hard cap velocity. Instead, it does not apply standard process to increase velocity in a direction where it is greater that the max. This allows outside forces to  push player velocity above its standard velocity max. For example, you can run to your max speed, but dash faster or take knockback to travel faster.

After brainstorming, the vertical velocity needs refactor to support natural and external velocities to allow gravity to always apply.

The dash doesn't feel great horizontally, because the drag kills it too quickly. This should be fixed by extending the dash state for the length of an animation and preventing the drag from being applied for the majority of the animation.