pixel movement feels right
Finally got the pixel movement to feel smooth and roll. I had to change the scale of the sprite to an integer, not a float, so it wouldn't do subpixel movement. I also changed the pixel rounding to round correctly instead of just flooring the position to the nearest pixel.