Completed draft of animation updater tool
Going forward, there will be tool scripts that edit game aspects in the editor so version control is extremely important.

This tool extracts sprite animations from the animated sprite 2d and moves them to the animation player so it is easy to so sprite animations when animating hitboxes and the like.

This tool script may not function yet