Added todo.txt and modified animation updater
This is a checkpoint! The changes to the animation updater may not behave as expected. Return to this commit if it is broken.

Modified the animation updater to only remove anim tracks that are being overwritten. So, you can change the frames of the sprite's animation after already setting parts of the animation player animation.