Started node flipper
While working on the node flipper, I realized I need a quick way to get references to all objects of a given their type. So the node flipper did not gain much progress.

The reference server has an empty function that is supposed to set a group for each node depending on its type so they are easy to find.