Began input refactor
Added Match select scene, where players press a to join and b to opt out. Then when there is at least 1 player, press start. This sends you to the play scene. The play scene now populates based on the player manager data. Since there is only one player, it defaults to selecting that player for the players, but when players do choose the character, it will be stored to the player profiles in the player manager.

Many signals now send with the player number instead of the input device. Keyboard is now considered a separate input player from the controller even if they are the same input device number.