Changed character tick process
*Still entirely broken

Player nodes now have process_tick function which is the rollback equivalent to physics_process.

The base of the player should be able to call it on each child node that has it. The Tick manager calls it on each player.

Need to look into extracting and loading player state.