not entirely working right now- Check this commit if inputs broken
First pass of Tick Manager complete. Shape still render in the editor correctly. Now to sync it with other character aspects and controller systems.

Fixed bug where game would crash if there were no players. This was a band aid fix- haven't looked at the code for the input manager in a while so hopefully this doesn't break it.

NEW TODO ITEMS FOR PHYSICS:
- Support objects changing sizes
- overlap for objects that aren't moving